python-workspace/apps/server/pyproject.toml
[project]
name = "blog-agent-server"
version = "0.1.0"
description = "Blog Agent gRPC Server with LlamaIndex workflows"
requires-python = ">=3.11"
dependencies = [
    "llama-index>=0.10.0",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "asyncpg>=0.29.0",
    "pgvector>=0.3.0",
    "tavily-python>=0.3.0",
    "pydantic>=2.5.0",
    "structlog>=24.1.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
    "black>=23.12.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

